fix: lodash breaking middleware
Importing lodash into middleware causes builds to fail because it's not
edge compatible (even though we don't use edge).
